+++ OIViewer/win32/MonitorInfo.h
@@ -22,9 +22,9 @@ namespace OIV
         MonitorInfo();
         void Refresh();
 
-        const MonitorDesc& getMonitorInfo(unsigned short monitorIndex, bool allowRefresh = false);
+        const MonitorDesc& getMonitorInfo(size_t monitorIndex, bool allowRefresh = false);
         const MonitorDesc& getMonitorInfo(HMONITOR hMonitor, bool allowRefresh = false);
-        const unsigned short getMonitorsCount() const;
+        const size_t getMonitorsCount() const;
         RECT getBoundingMonitorArea();
     private:
         using MapHMonitorToDesc = std::map<HMONITOR, MonitorDesc>;
